+++ src/components/textarea/textarea.js
@@ -58,12 +58,22 @@ class Textarea extends React.Component {
      * @type {Boolean}
      * @default true
      */
-    enforceCharacterLimit: PropTypes.bool
+    enforceCharacterLimit: PropTypes.bool,
+
+    /**
+     * Whether to display the character count message in red
+     *
+     * @property warnOverLimit
+     * @type {Boolean}
+     * @default false
+     */
+    warnOverLimit: PropTypes.bool
   }
 
   static defaultProps = {
     expandable: false,
-    enforceCharacterLimit: true
+    enforceCharacterLimit: true,
+    warnOverLimit: false
   }
 
   /**
@@ -154,6 +164,29 @@ class Textarea extends React.Component {
   }
 
   /**
+   * Uses the textAreaClasses method to add additional classes.
+   *
+   * @method textAreaClasses
+   * @return {String} textarea className
+   */
+  get textAreaClasses() {
+    return classNames (
+      'carbon-textarea__character-limit',
+      { 'over-limit': this.props.warnOverLimit && this.overLimit }
+    );
+  }
+
+  /**
+   * Returns if the character count exceeds the max
+   *
+   * @method overLimit
+   * @return {Boolean}
+   */
+  get overLimit() {
+    return calculateCharacterCount(this.props.value) > parseInt(this.props.characterLimit);
+  }
+
+  /**
    * A getter that combines props passed down from the input decorator with
    * textbox specific props.
    *
@@ -202,7 +235,7 @@ class Textarea extends React.Component {
   get characterCount() {
     if (this.props.characterLimit) {
       return (
-        <div className='carbon-textarea__character-limit' data-element='character-limit'>
+        <div className={ this.textAreaClasses } data-element='character-limit'>
           { I18n.t('textarea.limit.prefix', { defaultValue: 'You have used ' } ) }
           <span className='carbon-textarea__limit-used'>
             { I18n.toNumber(calculateCharacterCount(this.props.value), this.i18nNumberOpts) }